refactor: update routing to use 'home' instead of 'dashboard' and place login and setup under Auth layout
