+++ .trunk/trunk.yaml
@@ -18,25 +18,25 @@ runtimes:
 # This is the section where you manage your linters. (https://docs.trunk.io/check/configuration)
 lint:
   enabled:
-    - markdownlint@0.42.0
-    - eslint@9.14.0
+    - markdownlint@0.43.0
+    - eslint@9.16.0
     - actionlint@1.7.4
-    - bandit@1.7.10
+    - bandit@1.8.0
     - black@24.10.0
-    - checkov@3.2.296
+    - checkov@3.2.328
     - git-diff-check
     - isort@5.13.2
     - mypy@1.13.0
     - osv-scanner@1.9.1
-    - oxipng@9.1.2
-    - prettier@3.3.3
-    - ruff@0.7.3
+    - oxipng@9.1.3
+    - prettier@3.4.2
+    - ruff@0.8.1
     - shellcheck@0.10.0
     - shfmt@3.6.0
     - svgo@3.3.2
     - taplo@0.9.3
     - trivy@0.56.2
-    - trufflehog@3.83.6
+    - trufflehog@3.84.2
     - yamllint@1.35.1
   ignore:
     - linters: [ALL]
