+++ frontend/src/components/Home/Platforms.vue
