+++ frontend/src/components/Home/Stats.vue
