+++ frontend/src/main.ts
@@ -5,7 +5,7 @@ import "@/styles/scrollbar.css";
 import type { Events } from "@/types/emitter";
 import mitt from "mitt";
 import { createApp } from "vue";
-import App from "./RomM.vue";
+import App from "@/RomM.vue";
 
 const emitter = mitt<Events>();
 const app = createApp(App);
