+++ frontend/src/views/Player/EmulatorJS/Player.vue
